# server.js
const fs = require('fs')
const http = require('http')
const cors = require('cors')
const express = require('express')
const { Server } = require('socket.io')
const logger = require('./Logging')
const swaggerUi = require('swagger-ui-express')
const { ProcessingInstruction, ProcessingSegments, ProcessingSegmentsAsync } = require('./index')
const { createBullBoard } = require('@bull-board/api')
const { BullMQAdapter } = require('@bull-board/api/bullMQAdapter')
const { ExpressAdapter } = require('@bull-board/express')
const { segmentsQueue, computeStatisticsQueue, segmentsQueueEvents, addComputeStatisticsJob, computeStatisticsQueueEvents } = require('./workers')

const app = express()
const server = http.createServer(app)
const io = new Server(server, { cors: { origin: '*' } })

const port = process.env.PORT || 3001

io.on('connection', (socket) => {
  logger.info(`WebSocket connected: ${socket.id}`)
  socket.on('disconnect', () => {
    logger.info(`WebSocket disconnected: ${socket.id}`)
  })
})

segmentsQueueEvents.on('completed', ({ jobId, returnvalue }) => {
  try {
    const parsedResult = typeof returnvalue === 'string' ? JSON.parse(returnvalue) : returnvalue
    io.emit('process-completed', { jobId, result: parsedResult })
  } catch (err) {
    logger.error(`Error parsing process-completed result for job ${jobId}: ${err.message}`)
    io.emit('process-completed', { jobId, result: returnvalue })
  }
})

segmentsQueueEvents.on('failed', ({ jobId, failedReason }) => {
  io.emit('process-failed', { jobId, error: failedReason })
})

computeStatisticsQueueEvents.on('completed', ({ jobId, returnvalue }) => {
  try {
    const parsedResult = typeof returnvalue === 'string' ? JSON.parse(returnvalue) : returnvalue
    io.emit('compute-statistics-completed', { jobId, result: parsedResult })
  } catch (err) {
    logger.error(`Error parsing compute-statistics-completed result for job ${jobId}: ${err.message}`)
    io.emit('compute-statistics-completed', { jobId, result: returnvalue })
  }
})

computeStatisticsQueueEvents.on('failed', ({ jobId, failedReason }) => {
  io.emit('compute-statistics-failed', { jobId, error: failedReason })
})

app.use(cors())
app.use(express.json())

// Load generated swagger file
const swaggerFile = './swagger-output.json'
if (fs.existsSync(swaggerFile)) {
  const swaggerDocs = JSON.parse(fs.readFileSync(swaggerFile, 'utf-8'))
  app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocs))
} else {
  console.warn('Swagger file not found. Run "node swagger.js" first.')
}

// BullMQ Dashboard
const serverAdapter = new ExpressAdapter()
serverAdapter.setBasePath('/admin/queues')

createBullBoard({
  queues: [new BullMQAdapter(segmentsQueue), new BullMQAdapter(computeStatisticsQueue)],
  serverAdapter: serverAdapter,
})

app.use('/admin/queues', serverAdapter.getRouter())

app.post('/api/process', async (req, res) => {
  /* 
    #swagger.parameters['body'] = {
      in: 'body',
      description: 'Process a Bible command or segments',
      schema: { 
        type: 'object',
        properties: {
          segments: { 
            type: 'array',
            items: {
              type: 'object',
              properties: {
                book: { type: 'number' },
                chapter: { type: 'number' },
                from: { type: 'number' },
                to: { type: 'number' },
                publisher: { type: 'string' }
              }
            }
          },
          message: { type: 'string' }
        }
      }
    }
    #swagger.responses[200] = {
      description: 'Success',
      schema: { $ref: '#/definitions/ProcessResponse' }
    }
  */
  const { segments } = req.body
  logger.info(`Processing request: ${segments}`)
  try {
    if (segments && Array.isArray(segments)) {
      const result = await ProcessingSegmentsAsync(segments)
      if (result.error) {
        return res.status(400).json(result)
      }
      return res.json(result)
    }

    return res.status(400).json({ error: 'Segments (array) are required' })
  } catch (error) {
    console.error(error)
    res
      .status(500)
      .json({ error: 'Internal Server Error', details: error.message })
  }
})

app.post('/api/compute-statistics', async (req, res) => {
  /*
    #swagger.parameters['body'] = {
      in: 'body',
      description: 'Compute word-sequence statistics for gospel sections',
      schema: {
        type: 'object',
        properties: {
          verses: { type: 'array' },
          minLength: { type: 'number' },
          mode: { type: 'string' },
          similarityThreshold: { type: 'number' }
        }
      }
    }
    #swagger.responses[200] = {
      description: 'Job queued',
      schema: { jobId: 'string', status: 'queued' }
    }
  */
  const { verses, minLength, mode, similarityThreshold } = req.body
  logger.info(`Compute statistics request: ${verses?.length || 0} sections, mode=${mode}`)
  try {
    if (!verses || !Array.isArray(verses)) {
      return res.status(400).json({ error: 'verses (array) is required' })
    }
    const job = await addComputeStatisticsJob({ verses, minLength, mode, similarityThreshold })
    return res.json({ jobId: job.id, status: 'queued' })
  } catch (error) {
    console.error(error)
    res.status(500).json({ error: 'Internal Server Error', details: error.message })
  }
})

server.listen(port, () => {
  console.log(`Server is running on http://localhost:${port}`)
  console.log(`Swagger docs available at http://localhost:${port}/api-docs`)
  console.log(
    `BullMQ dashboard available at http://localhost:${port}/admin/queues`
  )
})